Move function to general helper class
